added strategies for spawning with different placements
