--- assets/scripts/system/WeaponSpawnerSystem.cs.uid
@@ -1,1 +1,0 @@
-uid://bfqc0lteae3ma
