--- assets/scripts/system/WeaponSpawningStrategy.cs
@@ -1,16 +1,0 @@
-namespace GunGame.assets.scripts.system
-{
-    public class WeaponSpawningStrategy
-    {
-    }
-
-    public enum StrategyType
-    {
-        TotallyRandom = 0,
-    }
-
-    public static class WeaponSpawningStrategyFactory
-    {
-
-    }
-}
