--- assets/scripts/system/WeaponSpawningStrategy.cs.uid
@@ -1,1 +1,0 @@
-uid://cl5vxetdrm1r0
