+++ scenes/main_scene.tscn
@@ -1,9 +1,10 @@
-[gd_scene load_steps=9 format=3 uid="uid://13hh6g4xgxgd"]
+[gd_scene load_steps=10 format=3 uid="uid://13hh6g4xgxgd"]
 
 [ext_resource type="SpriteFrames" uid="uid://c4cvgjgaobnej" path="res://scenes/kernel-animations.tres" id="1_w2twt"]
 [ext_resource type="Texture2D" uid="uid://bxodjufrjmmfa" path="res://assets/sprites/blocks/Blank-texture-64x64.png" id="3_7ykc4"]
 [ext_resource type="Script" uid="uid://knqkviihkbb8" path="res://assets/scripts/player/PlayerMovementRigidbody.cs" id="3_pu3yx"]
 [ext_resource type="PackedScene" uid="uid://dnok71q3gmhvn" path="res://scenes/system/weapon_spawner.tscn" id="4_pu3yx"]
+[ext_resource type="Script" uid="uid://bqtn8hp577gbo" path="res://assets/scripts/system/weapon_spawning/WeaponSpawnerSystem.cs" id="5_gubc5"]
 
 [sub_resource type="RectangleShape2D" id="RectangleShape2D_7ykc4"]
 size = Vector2(64, 64)
@@ -61,3 +62,4 @@ metadata/_edit_group_ = true
 shape = SubResource("CircleShape2D_gubc5")
 
 [node name="WeaponSpawner" parent="." instance=ExtResource("4_pu3yx")]
+script = ExtResource("5_gubc5")
