+++ scenes/system/weapon_spawner.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=2 format=3 uid="uid://dnok71q3gmhvn"]
 
-[ext_resource type="Script" uid="uid://bfqc0lteae3ma" path="res://assets/scripts/system/WeaponSpawnerSystem.cs" id="1_yx4eb"]
+[ext_resource type="Script" uid="uid://bqtn8hp577gbo" path="res://assets/scripts/system/weapon_spawning/WeaponSpawnerSystem.cs" id="1_yx4eb"]
 
 [node name="WeaponSpawner" type="Node2D"]
 script = ExtResource("1_yx4eb")
